requirements.txt
python-telegram-bot
requests
feedparser
sentence-transformers
scikit-learn
Commit changes